Horn melodies now display. Current build throws null error because mWeapon isn't initialized when calling mWeapon.getHornNotes.
